Dns: Implement name compression algorithm when writing RRs.
This compression algorithm is a trip. Basically, the message writer
needs to know when a QNAME needs to be written, because it needs to
look for the longest suffix that's already been written to the message
so that it can compress the name by only writing parts of the name
that have not before been seen in the message.